move concept api under transaction::concept
