+++ src/concept/thing.rs
@@ -20,11 +20,9 @@
  */
 
 use chrono::NaiveDateTime;
-use futures::Stream;
-use typeql_lang::pattern::Annotation;
 
-use super::{AttributeType, EntityType, RelationType, RoleType, ThingType};
-use crate::{common::IID, Result, Transaction};
+use super::{AttributeType, EntityType, RelationType};
+use crate::common::IID;
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum Thing {
@@ -51,128 +49,10 @@ pub struct Entity {
     pub type_: EntityType,
 }
 
-impl Entity {
-    pub fn new(iid: IID, type_: EntityType) -> Self {
-        Self { iid, type_ }
-    }
-
-    pub async fn is_deleted(&self, transaction: &Transaction<'_>) -> Result<bool> {
-        transaction.concept().get_entity(self.iid.clone()).await.map(|res| res.is_none())
-    }
-
-    pub async fn delete(&self, transaction: &Transaction<'_>) -> Result {
-        transaction.concept().thing_delete(Thing::Entity(self.clone())).await
-    }
-
-    pub fn get_has(
-        &self,
-        transaction: &Transaction<'_>,
-        attribute_types: Vec<AttributeType>,
-        annotations: Vec<Annotation>,
-    ) -> Result<impl Stream<Item = Result<Attribute>>> {
-        transaction.concept().thing_get_has(Thing::Entity(self.clone()), attribute_types, annotations)
-    }
-
-    pub async fn set_has(&self, transaction: &Transaction<'_>, attribute: Attribute) -> Result {
-        transaction.concept().thing_set_has(Thing::Entity(self.clone()), attribute).await
-    }
-
-    pub async fn unset_has(&self, transaction: &Transaction<'_>, attribute: Attribute) -> Result {
-        transaction.concept().thing_unset_has(Thing::Entity(self.clone()), attribute).await
-    }
-
-    pub fn get_relations(
-        &self,
-        transaction: &Transaction<'_>,
-        role_types: Vec<RoleType>,
-    ) -> Result<impl Stream<Item = Result<Relation>>> {
-        transaction.concept().thing_get_relations(Thing::Entity(self.clone()), role_types)
-    }
-
-    pub fn get_playing(&self, transaction: &Transaction<'_>) -> Result<impl Stream<Item = Result<RoleType>>> {
-        transaction.concept().thing_get_playing(Thing::Entity(self.clone()))
-    }
-}
-
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Relation {
     pub iid: IID,
     pub type_: RelationType,
-}
-
-impl Relation {
-    pub fn new(iid: IID, type_: RelationType) -> Self {
-        Self { iid, type_ }
-    }
-
-    pub async fn is_deleted(&self, transaction: &Transaction<'_>) -> Result<bool> {
-        transaction.concept().get_relation(self.iid.clone()).await.map(|res| res.is_none())
-    }
-
-    pub async fn delete(&self, transaction: &Transaction<'_>) -> Result {
-        transaction.concept().thing_delete(Thing::Relation(self.clone())).await
-    }
-
-    pub fn get_has(
-        &self,
-        transaction: &Transaction<'_>,
-        attribute_types: Vec<AttributeType>,
-        annotations: Vec<Annotation>,
-    ) -> Result<impl Stream<Item = Result<Attribute>>> {
-        transaction.concept().thing_get_has(Thing::Relation(self.clone()), attribute_types, annotations)
-    }
-
-    pub async fn set_has(&self, transaction: &Transaction<'_>, attribute: Attribute) -> Result {
-        transaction.concept().thing_set_has(Thing::Relation(self.clone()), attribute).await
-    }
-
-    pub async fn unset_has(&self, transaction: &Transaction<'_>, attribute: Attribute) -> Result {
-        transaction.concept().thing_unset_has(Thing::Relation(self.clone()), attribute).await
-    }
-
-    pub fn get_relations(
-        &self,
-        transaction: &Transaction<'_>,
-        role_types: Vec<RoleType>,
-    ) -> Result<impl Stream<Item = Result<Relation>>> {
-        transaction.concept().thing_get_relations(Thing::Relation(self.clone()), role_types)
-    }
-
-    pub fn get_playing(&self, transaction: &Transaction<'_>) -> Result<impl Stream<Item = Result<RoleType>>> {
-        transaction.concept().thing_get_playing(Thing::Relation(self.clone()))
-    }
-
-    pub async fn add_role_player(&self, transaction: &Transaction<'_>, role_type: RoleType, player: Thing) -> Result {
-        transaction.concept().relation_add_role_player(self.clone(), role_type, player).await
-    }
-
-    pub async fn remove_role_player(
-        &self,
-        transaction: &Transaction<'_>,
-        role_type: RoleType,
-        player: Thing,
-    ) -> Result {
-        transaction.concept().relation_remove_role_player(self.clone(), role_type, player).await
-    }
-
-    pub fn get_players_by_role_type(
-        &self,
-        transaction: &Transaction<'_>,
-        role_types: Vec<RoleType>,
-    ) -> Result<impl Stream<Item = Result<Thing>>> {
-        transaction.concept().relation_get_players_by_role_type(self.clone(), role_types)
-    }
-
-    pub fn get_role_players(
-        &self,
-        transaction: &Transaction<'_>,
-    ) -> Result<impl Stream<Item = Result<(RoleType, Thing)>>> {
-        transaction.concept().relation_get_role_players(self.clone())
-    }
-
-    pub fn get_relating(&self, transaction: &Transaction<'_>) -> Result<impl Stream<Item = Result<RoleType>>> {
-        transaction.concept().relation_get_relating(self.clone())
-    }
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -180,57 +60,6 @@ pub struct Attribute {
     pub iid: IID,
     pub type_: AttributeType,
     pub value: Value,
-}
-
-impl Attribute {
-    pub fn new(iid: IID, type_: AttributeType, value: Value) -> Self {
-        Self { iid, type_, value }
-    }
-
-    pub async fn is_deleted(&self, transaction: &Transaction<'_>) -> Result<bool> {
-        transaction.concept().get_attribute(self.iid.clone()).await.map(|res| res.is_none())
-    }
-
-    pub async fn delete(&self, transaction: &Transaction<'_>) -> Result {
-        transaction.concept().thing_delete(Thing::Attribute(self.clone())).await
-    }
-
-    pub fn get_has(
-        &self,
-        transaction: &Transaction<'_>,
-        attribute_types: Vec<AttributeType>,
-        annotations: Vec<Annotation>,
-    ) -> Result<impl Stream<Item = Result<Attribute>>> {
-        transaction.concept().thing_get_has(Thing::Attribute(self.clone()), attribute_types, annotations)
-    }
-
-    pub async fn set_has(&self, transaction: &Transaction<'_>, attribute: Attribute) -> Result {
-        transaction.concept().thing_set_has(Thing::Attribute(self.clone()), attribute).await
-    }
-
-    pub async fn unset_has(&self, transaction: &Transaction<'_>, attribute: Attribute) -> Result {
-        transaction.concept().thing_unset_has(Thing::Attribute(self.clone()), attribute).await
-    }
-
-    pub fn get_relations(
-        &self,
-        transaction: &Transaction<'_>,
-        role_types: Vec<RoleType>,
-    ) -> Result<impl Stream<Item = Result<Relation>>> {
-        transaction.concept().thing_get_relations(Thing::Attribute(self.clone()), role_types)
-    }
-
-    pub fn get_playing(&self, transaction: &Transaction<'_>) -> Result<impl Stream<Item = Result<RoleType>>> {
-        transaction.concept().thing_get_playing(Thing::Attribute(self.clone()))
-    }
-
-    pub fn get_owners(
-        &self,
-        transaction: &Transaction<'_>,
-        thing_type: Option<ThingType>,
-    ) -> Result<impl Stream<Item = Result<Thing>>> {
-        transaction.concept().attribute_get_owners(self.clone(), thing_type)
-    }
 }
 
 #[derive(Clone, Debug, PartialEq)]
